Move backend, database, and cache into server directory.
