+++ lsif/src/server/backend/cache.metrics.ts
