+++ lsif/src/server/backend/database.ts
@@ -1,13 +1,13 @@
 import * as lsp from 'vscode-languageserver-protocol'
 import { Connection } from 'typeorm'
 import { ConnectionCache, DocumentCache, EncodedJsonCacheValue, ResultChunkCache } from './cache'
-import { hashKey, mustGet } from './util'
-import { instrument } from './metrics'
-import { databaseQueryDurationHistogram, databaseQueryErrorsCounter } from './database.metrics'
-import { DefaultMap } from './default-map'
-import { gunzipJSON } from './encoding'
+import { hashKey, mustGet } from '../../util'
+import { instrument } from '../../metrics'
+import { databaseQueryDurationHistogram, databaseQueryErrorsCounter } from '../../database.metrics'
+import { DefaultMap } from '../../default-map'
+import { gunzipJSON } from '../../encoding'
 import { isEqual, uniqWith } from 'lodash'
-import { DumpId, LsifDump } from './xrepo.models'
+import { DumpId, LsifDump } from '../../xrepo.models'
 import {
     DefinitionModel,
     DocumentData,
@@ -20,9 +20,9 @@ import {
     ResultChunkModel,
     DocumentPathRangeId,
     entities,
-} from './database.models'
-import { DefinitionReferenceResultId, RangeId } from './database.types'
-import { TracingContext, logSpan } from './tracing'
+} from '../../database.models'
+import { DefinitionReferenceResultId, RangeId } from '../../database.types'
+import { TracingContext, logSpan } from '../../tracing'
 
 /**
  * A wrapper around operations for single repository/commit pair.
