+++ lsif/src/server/routes/lsif.ts
@@ -6,7 +6,7 @@ import express from 'express'
 import pTimeout from 'p-timeout'
 import uuid from 'uuid'
 import { addTags, logAndTraceCall, TracingContext } from '../../tracing'
-import { Backend, ReferencePaginationCursor } from '../../backend'
+import { Backend, ReferencePaginationCursor } from '../backend/backend'
 import { DEFAULT_REFERENCES_NUM_REMOTE_DUMPS, STORAGE_ROOT } from '../settings'
 import { encodeCursor, parseCursor } from '../pagination/cursor'
 import { enqueue } from '../../queue'
