+++ lsif/src/tests/util.ts
@@ -1,5 +1,5 @@
 import rmfr from 'rmfr'
-import { Backend, ReferencePaginationCursor } from '../backend'
+import { Backend, ReferencePaginationCursor } from '../server/backend/backend'
 import { Configuration } from '../config'
 import { createCleanPostgresDatabase, createStorageRoot, convertTestData } from '../test-utils'
 import { XrepoDatabase } from '../xrepo'
